pyproject.toml
[project]
name = "lightning-template-hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "black==24.8.0",
    "coverage==7.6.1",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "lightning[extra]==2.4.0",
    "loguru==0.7.2",
    "pytest==8.3.3",
    "rich==13.8.1",
    "rootutils==1.0.7",
    "tensorboard==2.17.1",
    "timm==1.0.9",
    "torch==2.4.1",
    "torchvision==0.19.1",
]

[tool.coverage.report]
include = ["src/**/*.py"]
show_missing = true
